Feat: Add login

+++ public/script.js
@@ -65,7 +65,7 @@ function renderHeroesListButton() {
   heroesListButtonContainer.append(listIcon, spinIcon, checkIcon);
 
   heroesListButtonContainer.addEventListener('click', () => {
-    alert();
+    renderLoginContainer();
   });
 
   const userSection = document.getElementById('userSectionContainer');
@@ -555,6 +555,12 @@ async function loadHeroes() {
 }
 
 async function saveHero() {
+  const token = localStorage.getItem('token');
+  if (!token) {
+    alert('User not logged.');
+    return;
+  }
+
   if (tempHero) {
     getInputs();
 
@@ -601,6 +607,12 @@ async function saveHero() {
 }
 
 async function updateHero(hero) {
+  const token = localStorage.getItem('token');
+  if (!token) {
+    alert('User not logged.');
+    return;
+  }
+
   try {
     const response = await fetch(`${BASE_URL}/heroes`, {
       method: 'PUT',
@@ -618,6 +630,12 @@ async function updateHero(hero) {
 }
 
 async function saveNewHero(hero) {
+  const token = localStorage.getItem('token');
+  if (!token) {
+    alert('User not logged.');
+    return;
+  }
+
   try {
     const response = await fetch(`${BASE_URL}/heroes`, {
       method: 'POST',
@@ -635,12 +653,18 @@ async function saveNewHero(hero) {
 }
 
 async function deleteHero() {
+  const token = localStorage.getItem('token');
+  if (!token) {
+    alert('User not logged.');
+    return;
+  }
   if (tempHero && tempHero.id) {
     try {
       const response = await fetch(`${BASE_URL}/heroes`, {
         method: 'DELETE',
         headers: {
           'Content-Type': 'application/json',
+          Authorization: `Bearer ${token}`,
         },
         body: JSON.stringify({ id: tempHero.id }),
       });
@@ -654,7 +678,103 @@ async function deleteHero() {
     }
   }
 }
+// 👇    👇    USER AUTH  👇    👇
+//
+//
+function renderLoginContainer() {
+  const loginContainer = document.getElementById('loginContainer');
 
+  if (loginContainer) {
+    loginContainer.remove();
+    return;
+  }
+
+  const token = localStorage.getItem('token');
+
+  const newLoginContainer = document.createElement('div');
+  newLoginContainer.className = 'login-container';
+  newLoginContainer.id = 'loginContainer';
+
+  const mainContainer = document.getElementById('mainContainer');
+  const inputContainer = document.createElement('div');
+  inputContainer.className = 'input-container';
+  const buttonsContainer = document.createElement('div');
+  buttonsContainer.className = 'buttons-container';
+  buttonsContainer.id = 'buttonsContainer';
+
+  mainContainer.append(newLoginContainer);
+  newLoginContainer.append(inputContainer, buttonsContainer);
+
+  if (!token) {
+    const loginInput = document.createElement('input');
+    loginInput.id = 'loginInput';
+    const passwordInput = document.createElement('input');
+    passwordInput.type = 'password';
+    passwordInput.id = 'passwordInput';
+    inputContainer.append(loginInput, passwordInput);
+
+    const loginButton = document.createElement('div');
+    loginButton.className = 'login-button';
+    loginButton.id = 'loginButton';
+    loginButton.textContent = 'Login';
+    loginButton.addEventListener('click', () => {
+      const username = document.getElementById('loginInput').value.toLowerCase();
+      const password = document.getElementById('passwordInput').value;
+
+      login(username, password);
+    });
+
+    const signInButton = document.createElement('div');
+    signInButton.className = 'signin-button';
+    signInButton.id = 'signInButton';
+    signInButton.textContent = 'Sign in';
+    signInButton.addEventListener('click', () => {
+      const username = document.getElementById('loginInput').value.toLowerCase();
+      const password = document.getElementById('passwordInput').value;
+
+      register(username, password);
+    });
+
+    buttonsContainer.append(loginButton, signInButton);
+  } else {
+    const userPicture = document.createElement('div');
+    userPicture.className = 'user-picture';
+    const myHeroesButton = document.createElement('div');
+    myHeroesButton.textContent = 'My Heroes';
+    inputContainer.appendChild(userPicture);
+    buttonsContainer.appendChild(myHeroesButton);
+  }
+}
+
+async function login(username, password) {
+  const response = await fetch(`${BASE_URL}/users/login`, {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ username, password }),
+  });
+  if (response.ok) {
+    const { token } = await response.json();
+    localStorage.setItem('token', token);
+    alert('User logged in');
+    renderLoginContainer();
+  } else {
+    alert('Login failed');
+  }
+}
+
+async function register(username, password) {
+  const response = await fetch(`${BASE_URL}/users/register`, {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ username, password }),
+  });
+  if (response.ok) {
+    alert('User registered successfully');
+    renderLoginContainer();
+  } else {
+    alert('Registration failed');
+  }
+}
 // 👇    👇    UTILITY   👇    👇
 //
 //
